+++ metatensor-core/tests/cpp/data.cpp
@@ -79,7 +79,10 @@ TEST_CASE("Data Array") {
 
         uintptr_t new_shape[] = {1, 2, 3, 4};
         shape_count = 4;
-        status = array.create(array.ptr, new_shape, shape_count, &new_array);
+        auto fill_value_2 = DataArrayBase::to_mts_array_t(
+            std::make_unique<SimpleDataArray<double>>(std::vector<uintptr_t>{1}, 0.0)
+        );
+        status = array.create(array.ptr, new_shape, shape_count, fill_value_2, &new_array);
         CHECK(status == MTS_SUCCESS);
 
         status = new_array.shape(new_array.ptr, &shape, &shape_count);
@@ -91,6 +94,7 @@ TEST_CASE("Data Array") {
         CHECK(shape[2] == 3);
         CHECK(shape[3] == 4);
         new_array.destroy(new_array.ptr);
+        fill_value_2.destroy(fill_value_2.ptr);
     }
 
     array.destroy(array.ptr);