+++ python/metatensor_core/metatensor/_c_api.py
@@ -123,7 +123,7 @@ mts_array_t._fields_ = [
     ("shape", CFUNCTYPE(mts_status_t, ctypes.c_void_p, POINTER(POINTER(c_uintptr_t)), POINTER(c_uintptr_t))),
     ("reshape", CFUNCTYPE(mts_status_t, ctypes.c_void_p, POINTER(c_uintptr_t), c_uintptr_t)),
     ("swap_axes", CFUNCTYPE(mts_status_t, ctypes.c_void_p, c_uintptr_t, c_uintptr_t)),
-    ("create", CFUNCTYPE(mts_status_t, ctypes.c_void_p, POINTER(c_uintptr_t), c_uintptr_t, POINTER(mts_array_t))),
+    ("create", CFUNCTYPE(mts_status_t, ctypes.c_void_p, POINTER(c_uintptr_t), c_uintptr_t, mts_array_t, POINTER(mts_array_t))),
     ("copy", CFUNCTYPE(mts_status_t, ctypes.c_void_p, POINTER(mts_array_t))),
     ("destroy", CFUNCTYPE(None, ctypes.c_void_p)),
     ("move_data", CFUNCTYPE(mts_status_t, ctypes.c_void_p, ctypes.c_void_p, POINTER(mts_data_movement_t), c_uintptr_t)),
@@ -349,6 +349,7 @@ def setup_functions(lib):
         POINTER(mts_tensormap_t),
         mts_labels_t,
         ctypes.c_bool,
+        mts_array_t,
     ]
     lib.mts_tensormap_keys_to_properties.restype = POINTER(mts_tensormap_t)
 
@@ -363,6 +364,7 @@ def setup_functions(lib):
         POINTER(mts_tensormap_t),
         mts_labels_t,
         ctypes.c_bool,
+        mts_array_t,
     ]
     lib.mts_tensormap_keys_to_samples.restype = POINTER(mts_tensormap_t)