+++ rust/metatensor/src/data/mod.rs
@@ -38,7 +38,9 @@ mod tests {
         assert_eq!(get_data_origin(array.as_raw().origin().unwrap()).unwrap(), "rust.Box<dyn Array>");
         assert_eq!(array.as_array(), ArcArray::from_elem(vec![4, 2], 1.0));
 
-        let other = unsafe { ArrayRef::from_raw(array.as_raw().create(&[5, 3, 7, 12]).unwrap()) };
+        let fv_data: Box<dyn Array> = Box::new(ArcArray::from_elem(vec![1], 0.0));
+        let fv = crate::c_api::mts_array_t::from(fv_data);
+        let other = unsafe { ArrayRef::from_raw(array.as_raw().create(&[5, 3, 7, 12], &fv).unwrap()) };
         assert_eq!(other.as_raw().shape().unwrap(), [5, 3, 7, 12]);
         assert_eq!(get_data_origin(other.as_raw().origin().unwrap()).unwrap(), "rust.Box<dyn Array>");
         assert_eq!(other.as_array(), ArcArray::from_elem(vec![5, 3, 7, 12], 0.0));
@@ -49,9 +51,10 @@ mod tests {
         let array = ArcArray::from_elem(vec![3, 2, 2, 4], 1.0);
         let array = unsafe { ArrayRefMut::new((Box::new(array) as Box<dyn Array>).into()) };
 
-        let mut other = unsafe { ArrayRefMut::new(array.as_raw().create(&[1, 2, 2, 8]).unwrap()) };
-        let expected = ArcArray::from_elem(vec![1, 2, 2, 8], 0.0);
-        assert_eq!(other.as_array(), expected);
+        let fv_data: Box<dyn Array> = Box::new(ArcArray::from_elem(vec![1], 0.0));
+        let fv = crate::c_api::mts_array_t::from(fv_data);
+        let mut other = unsafe { ArrayRefMut::new(array.as_raw().create(&[1, 2, 2, 8], &fv).unwrap()) };
+        assert_eq!(other.as_array(), ArcArray::from_elem(vec![1, 2, 2, 8], 0.0));
 
         let mapping = mts_data_movement_t {
             sample_in: 1,